+++ Software/MATLAB/SQL/app/system/+azure/+documentdb/@Document/toJson.m
@@ -1,0 +1,8 @@
+function result = toJson(obj)
+% TOJSON Converts to a JSON string in character vector form
+
+% Copyright 2019 The MathWorks, Inc.
+
+result = char(obj.Handle.toJson());
+
+end