+++ Software/MATLAB/SQL/app/system/+azure/+documentdb/@Document/toString.m
@@ -1,0 +1,8 @@
+function result = toString(obj)
+% TOSTRING Gets string representation of property bag in character vector form
+
+% Copyright 2019 The MathWorks, Inc.
+
+result = char(obj.Handle.toString());
+
+end