+++ RELEASENOTES.md
@@ -1,16 +1,27 @@
 # MATLAB Interface *for Azure Cosmos DB*
 # Release Notes
 
-## Release 0.1.2 April 18 2019
+## Release 0.1.3 (14th March 2020)
+* Fixed missing network preferences image
+* Added Document getBoolean, toString() and toJson() methods
+* Clarified Database Toolbox requirement
+* Improved network configuration proxy documentation
+* Added SQL API proxy support
+* Added ConsistencyLevel support
+* Added ConnectionMode enum
+* Added Security notice
+* Updated License terms and file format
+
+## Release 0.1.2 (April 18th 2019)
 * Added SQL API support
 * Improved documentation
 
-## Release 0.1.1 October 31 2018
+## Release 0.1.1 (October 31st 2018)
 * Improved documentation
 * Reorganized content
 * Interim external release of MongoDB support
 
-## Release 0.1.0 July 13 2018
+## Release 0.1.0 (July 13th 2018)
 * *Internal release only*
 * Support for Cosmos DB
     * SQL API